refactor(68.2-06): relocate pure version-transform helpers
- Create apps/web/src/lib/version-transforms.ts with shouldCreateVersion,
  computeRestoreVersionUpdate, computeDeleteVersionUpdate relocated verbatim
  from file-metadata.service.ts (RESEARCH Pitfall 4 -- these are pure
  write-path helpers, not part of the read chain being deleted in Plan 11)
- Repoint useFileVersions.ts, useFileOperations.ts, TextEditorDialog.tsx to
  import the three helpers from lib/version-transforms instead of the
  file-metadata service
- resolveFileMetadata stays in file-metadata.service.ts for now (Task 2
  replaces its call sites; the file itself is deleted in Plan 11)

Entire-Checkpoint: 53691e671622
