+++ apps/web/src/components/file-browser/TextEditorDialog.tsx
@@ -8,7 +8,8 @@ import { useVaultStore } from '../../stores/vault.store';
 import { getRootFolderState } from '../../hooks/folder-helpers';
 import { runWithFailureUx } from '../../hooks/useMutationFailureUx';
 import { downloadFileFromIpns, triggerBrowserDownload } from '../../services/download.service';
-import { resolveFileMetadata, shouldCreateVersion } from '../../services/file-metadata.service';
+import { resolveFileMetadata } from '../../services/file-metadata.service';
+import { shouldCreateVersion } from '../../lib/version-transforms';
 import { addToIpfs, unpinFromIpfs } from '../../lib/api/ipfs';
 import { getSdkClient, hasSdkClient } from '../../lib/sdk-provider';
 import { logger } from '../../lib/logger';
