+++ apps/web/src/services/file-metadata.service.ts
@@ -12,31 +12,18 @@
  * anti-rollback floor sdk-core's raw resolveIpnsRecord does not) rather than the
  * sdk-core ctx-based helpers, since the web app has no `SdkContext` bridge.
  *
- * `shouldCreateVersion` / `computeRestoreVersionUpdate` / `computeDeleteVersionUpdate`
- * are pure, no-publish transforms over `NodeContent.versions` (68.1-12) — ported from
- * the pre-v3 legacy implementation (commit b24e78e90) and adapted to node/v3's raw
- * `Uint8Array` fileKey / base64 fileIv contract (NODE-02).
+ * `resolveFileMetadata` is a deletion target for 68.2-11 (SC#1 read-chain
+ * collapse) — its callers are rewired in 68.2-06/07/08 onto the SDK facade. The
+ * three pure version-transform helpers that used to live in this file
+ * (`shouldCreateVersion`/`computeRestoreVersionUpdate`/`computeDeleteVersionUpdate`)
+ * were relocated to `../lib/version-transforms.ts` in 68.2-06 (RESEARCH Pitfall 4)
+ * — they are NOT part of the read chain and must survive this file's deletion.
  */
 
-import type { SealedChildRef, VersionEntry, NodeContent, PublishedNode } from '@cipherbox/core';
+import type { SealedChildRef, NodeContent, PublishedNode } from '@cipherbox/core';
 import { unsealNode, unsealChildReadKey } from '@cipherbox/core';
-import { type UpdateFileContentParams } from '@cipherbox/sdk-core';
 import { resolveIpnsRecord } from './ipns.service';
 import { fetchFromIpfs } from '../lib/api/ipfs';
-import { useVaultSettingsStore } from '../stores/vault-settings.store';
-
-/** Content fields applied as `updates` when routing a file metadata write through the SDK client. */
-export type FileContentUpdates = UpdateFileContentParams;
-
-/** Decodes a base64 string to a Uint8Array (VersionEntry.fileIv is base64, v3 contract). */
-function base64ToBytes(b64: string): Uint8Array {
-  const bin = atob(b64);
-  const bytes = new Uint8Array(bin.length);
-  for (let i = 0; i < bin.length; i++) {
-    bytes[i] = bin.charCodeAt(i);
-  }
-  return bytes;
-}
 
 /**
  * Resolve a file Node's content.
@@ -97,96 +84,4 @@ export async function resolveFileMetadata(
     fileReadKey.fill(0);
     fileReadKey = null;
   }
-}
-
-/**
- * Determine whether a file content update should create a new version entry.
- *
- * Ported from the pre-v3 policy (commit b24e78e90), adapted to `VersionEntry`'s
- * `createdAt` field (NODE-02):
- * - `forceVersion` (explicit re-upload, e.g. `ReplaceFileDialog`) always versions.
- * - No existing versions → always version (first version always created).
- * - Otherwise version only if the newest entry is older than the user-configurable
- *   cooldown (`vault-settings.store`'s `versionCooldownMinutes`).
- *
- * @param currentVersions - Current file's version history (may be empty/undefined)
- * @param forceVersion - Whether to force version creation regardless of cooldown
- */
-export function shouldCreateVersion(
-  currentVersions: VersionEntry[] | undefined,
-  forceVersion: boolean
-): boolean {
-  if (forceVersion) return true;
-  if (!currentVersions || currentVersions.length === 0) return true;
-
-  const cooldownMs = useVaultSettingsStore.getState().settings.versionCooldownMinutes * 60 * 1000;
-  const newestCreatedAt = currentVersions[0].createdAt;
-  return Date.now() - newestCreatedAt >= cooldownMs;
-}
-
-/**
- * Compute the metadata transform for restoring a previous version (pure, no publish).
- *
- * Picks `versions[versionIndex]` as the new live content (`updates`, decoding its
- * base64 `fileIv` back to raw bytes for `UpdateFileContentParams`) and removes it
- * from the retained history (`retainedVersions`) so it is not duplicated once the
- * caller republishes. The caller is expected to pass `createVersion: true` to
- * `client.restoreFileVersion` so sdk-core's `updateFileMetadata` folds the
- * pre-restore live content into `retainedVersions` itself (capped, pruning as
- * needed) — this function never needs to compute that fold or cap directly, since
- * it only has `versions`, not the live content descriptor.
- *
- * @param versions - Current file's version history
- * @param versionIndex - Index of the version to restore (0 = newest past version)
- */
-export function computeRestoreVersionUpdate(
-  versions: VersionEntry[],
-  versionIndex: number
-): {
-  updates: Omit<FileContentUpdates, 'mimeType'>;
-  retainedVersions: VersionEntry[];
-  prunedCids: string[];
-} {
-  if (versionIndex < 0 || versionIndex >= versions.length) {
-    throw new Error('Invalid version index');
-  }
-
-  const versionToRestore = versions[versionIndex];
-  const retainedVersions = versions.filter((_, i) => i !== versionIndex);
-
-  return {
-    updates: {
-      cid: versionToRestore.cid,
-      fileKey: versionToRestore.fileKey,
-      fileIv: base64ToBytes(versionToRestore.fileIv),
-      size: versionToRestore.size,
-      encryptionMode: versionToRestore.encryptionMode,
-    },
-    retainedVersions,
-    prunedCids: [],
-  };
-}
-
-/**
- * Compute the metadata transform for deleting a past version (pure, no publish).
- *
- * Removes the version at `versionIndex` from the history. The live content is
- * untouched — the caller publishes the same content with a pruned version list
- * (`createVersion: false`).
- *
- * @param versions - Current file's version history
- * @param versionIndex - Index of the version to delete
- */
-export function computeDeleteVersionUpdate(
-  versions: VersionEntry[],
-  versionIndex: number
-): { retainedVersions: VersionEntry[]; deletedCid: string } {
-  if (versionIndex < 0 || versionIndex >= versions.length) {
-    throw new Error('Invalid version index');
-  }
-
-  const deletedCid = versions[versionIndex].cid;
-  const retainedVersions = versions.filter((_, i) => i !== versionIndex);
-
-  return { retainedVersions, deletedCid };
 }
